% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{annualAnomaly}
\alias{annualAnomaly}
\title{Calculate annual anomaly values}
\usage{
annualAnomaly(d, anomaly)
}
\arguments{
\item{d}{a data.frame containing year, month, and at least anomaly, and optionally normalizedAnomaly}

\item{anomaly}{a character string naming the column of monthly anomaly which should be aggregated into annual anomaly}
}
\value{
a data.frame with year, anomaly, and optionally normalizedAnomaly
}
\description{
This function calculates annual anomaly values from provided
monthly anomaly values.
}
\author{
Chantelle Layton
}
