% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{annualAnomaly_byRegion}
\alias{annualAnomaly_byRegion}
\title{Annual Anomalies by region}
\usage{
annualAnomaly_byRegion(data, climatologyYears)
}
\arguments{
\item{data}{A data frame of annual data containing AT LEAST metadata columns,
year and section or station as well as one or more columns of data. If
gathering data from \code{azmpdata} package, data frame should be \code{...Annual_Section} or
\code{...Annual_Station} format.}

\item{climatologyYears}{A vector of two numbers indicating the range to be used to calculate climatology (eg. c(1999, 2015))}
}
\value{
A data frame with annual anomalies named by the variable (NOTE: the anomalies are not named with 'anomaly', be careful of confusing with actual data values)
}
\description{
This function caluclates annual anomalies ONLY FROM annual data. It can
accept annual data from multiple stations or sections in the same dataframe
although users should be cautious that sections/ stations are properly
labeled.
}
