% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{annualNormalizedAnomaly}
\alias{annualNormalizedAnomaly}
\title{Calculate normalized annual anomaly values}
\usage{
annualNormalizedAnomaly(d, climatologyYears)
}
\arguments{
\item{d}{a data.frame containing year, month, and at least anomaly}

\item{climatologyYears}{a vector of length two indicating the range of years
to calculate the climatology.}
}
\value{
a data.frame with year, normalizedAnomaly, and other columns in d
}
\description{
This function calculates normalized annual anomaly values from provided
monthly anomaly values.
}
\author{
Chantelle Layton
}
