% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_spec.R
\name{build_column_spec}
\alias{build_column_spec}
\title{Build Column Specification}
\usage{
build_column_spec(
  k,
  d,
  tvp,
  common_effects,
  subgroup,
  common_tvp_effects,
  num_periods,
  num_subgroups
)
}
\description{
Internal function to compute column index ranges.
}
\keyword{internal}
