% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_initial_coefs.R
\name{build_cv_folds}
\alias{build_cv_folds}
\title{Build CV fold structure for blocked time series cross-validation}
\usage{
build_cv_folds(Ak, nfolds)
}
\arguments{
\item{Ak}{List of design matrices per subject}

\item{nfolds}{Number of folds}
}
\value{
List of fold ID vectors per subject
}
\description{
Build CV fold structure for blocked time series cross-validation
}
\keyword{internal}
