% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_initial_coefs.R
\name{build_period_folds}
\alias{build_period_folds}
\title{Build CV folds for a single period}
\usage{
build_period_folds(idx, nfolds)
}
\arguments{
\item{idx}{Indices within the period}

\item{nfolds}{Number of folds}
}
\value{
Fold ID vector
}
\description{
Build CV folds for a single period
}
\keyword{internal}
