% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_spec.R
\name{build_row_spec}
\alias{build_row_spec}
\title{Build Row Specification}
\usage{
build_row_spec(k, n, tvp, breaks)
}
\arguments{
\item{n}{Integer vector. Number of rows per subject (already accounts for lag).
This should be ntk from constructModel, which is nrow(dat$b).}
}
\description{
Internal function to compute row index ranges.
}
\keyword{internal}
