% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maity_debiasing.R
\name{build_theta_nodewise}
\alias{build_theta_nodewise}
\title{Build nodewise LASSO precision estimate}
\usage{
build_theta_nodewise(X, nfolds = 5)
}
\arguments{
\item{X}{Design matrix (n x d)}

\item{nfolds}{Number of CV folds}
}
\value{
d x d precision estimate (symmetrized)
}
\description{
For each column j of X, regress X_j on X_{-j} via LASSO.
Construct the precision matrix Omega from the nodewise coefficients
and residual variances, then symmetrize.
}
\keyword{internal}
