% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.mvnorm.plot.R
\name{check.mvnorm.plot}
\alias{check.mvnorm.plot}
\title{Checks for multivariate normality in data}
\usage{
check.mvnorm.plot(data)
}
\arguments{
\item{data}{A dataframe with multiple quantitative variables}
}
\value{
A qq-plot for multivariate normality
}
\description{
Checks for multivariate normality in data
}
\examples{
check.mvnorm.plot(iris[,1:4])

}
