% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_initial_coefs.R
\name{compute_common_effects}
\alias{compute_common_effects}
\title{Compute common effects as median across subjects (or periods for k=1 TVP)}
\usage{
compute_common_effects(total, periods, k, d, tvp, include_common)
}
\description{
Compute common effects as median across subjects (or periods for k=1 TVP)
}
\keyword{internal}
