% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_initial_coefs.R
\name{compute_subgroup_effects}
\alias{compute_subgroup_effects}
\title{Compute subgroup effects as median within subgroup minus common}
\usage{
compute_subgroup_effects(total, subgroup_membership, common, d)
}
\description{
Compute subgroup effects as median within subgroup minus common
}
\keyword{internal}
