% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maity_debiasing.R
\name{debias_var1_single}
\alias{debias_var1_single}
\title{Debias a single subject's VAR(1) coefficient matrix}
\usage{
debias_var1_single(X, Y, Phi_hat, nfolds = 5)
}
\arguments{
\item{X}{Design matrix (n x d)}

\item{Y}{Response matrix (n x d)}

\item{Phi_hat}{LASSO estimate (d x d, row i = equation i)}

\item{nfolds}{Number of CV folds}
}
\value{
Debiased d x d coefficient matrix
}
\description{
Debias a single subject's VAR(1) coefficient matrix
}
\keyword{internal}
