% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_initial_coefs.R
\name{decompose_tvp_effects}
\alias{decompose_tvp_effects}
\title{Decompose TVP effects into common_tvp and unique tvp components}
\usage{
decompose_tvp_effects(
  periods,
  k,
  d,
  common,
  include_common,
  include_common_tvp
)
}
\description{
Decompose TVP effects into common_tvp and unique tvp components
}
\keyword{internal}
