% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_initial_coefs.R
\name{estimate_glmnet}
\alias{estimate_glmnet}
\title{Estimate coefficients using glmnet (lasso or ridge)}
\usage{
estimate_glmnet(A, b, alpha, folds, lambda_choice, glmnet_intercept = FALSE)
}
\arguments{
\item{A}{Design matrix (n x d)}

\item{b}{Response matrix (n x d)}

\item{alpha}{Elastic net mixing parameter (1 = lasso, 0 = ridge)}

\item{folds}{Fold ID vector for CV}

\item{lambda_choice}{Which lambda to use ("lambda.min" or "lambda.1se")}

\item{glmnet_intercept}{Logical. Whether glmnet should fit an intercept (default FALSE)}
}
\value{
Coefficient matrix (d x d)
}
\description{
Estimate coefficients using glmnet (lasso or ridge)
}
\keyword{internal}
