% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_initial_coefs.R
\name{estimate_initial_coefs}
\alias{estimate_initial_coefs}
\title{Estimate consistent first-stage coefficients}
\usage{
estimate_initial_coefs(
  Ak,
  bk,
  d,
  k,
  lassotype,
  weightest,
  subgroup_membership,
  subgroup,
  nlambda1,
  nlambda2
)
}
\description{
`estimate_initial_coefs` returns consistent first-stage estimates to be used
for calculating adaptive weights used in regularization.
}
\details{
This function returns a list containing the following elements:
* common_effects: a d x d matrix containing consistent esimtates of the common 
  effects
* subgroup_effects: a list of length max(subgroup), where each element is a d x d matrix 
  containing consistent esimtates of the subgroup effects
* unique_effects: a list of length k, where each element is a d x d matrix 
  containing consistent esimtates of the unique effects
* total_effects: a list of length k, where each element is a d x d matrix 
  containing consistent esimtates of the total effects

#' @export
}
