% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_initial_coefs.R
\name{estimate_ols}
\alias{estimate_ols}
\title{Estimate coefficients using OLS}
\usage{
estimate_ols(A, b)
}
\arguments{
\item{A}{Design matrix (n x d)}

\item{b}{Response matrix (n x d)}
}
\value{
Coefficient matrix (d x d)
}
\description{
Estimate coefficients using OLS
}
\keyword{internal}
