% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_multivar_performance_old.R
\name{eval_multivar_performance_old}
\alias{eval_multivar_performance_old}
\title{Summarize multivar performance against simulation truth (robust to missing parts)}
\usage{
eval_multivar_performance_old(
  sim_obj,
  fit_obj,
  eps = 1e-08,
  reduced.output = FALSE
)
}
\description{
Compare estimated transition matrices to simulation truth, but only for
components that actually exist in both `sim_obj` and `fit_obj`. For example,
if `fit_obj$mats` only contains `total`, only "total" rows are returned.
}
