% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_multivar_hyperparams.R
\name{extract_multivar_hyperparams}
\alias{extract_multivar_hyperparams}
\title{Extract hyperparameter info from multivar CV fit}
\usage{
extract_multivar_hyperparams(object, fit)
}
\arguments{
\item{object}{multivar model object used to call cv_multivar(...)}

\item{fit}{result returned by cv_multivar(...); MSFE is assumed to be fit[[2]]}
}
\description{
Run this right after:
  fit <- cv_multivar(...)
}
\details{
It reconstructs:
  - full lambda1 grid (from the model object)
  - implied lambda2 grid (lambda1 * ratio)
  - scenario table (one row per scenario)
  - MSFE surface (nlam x n_scenarios) from fit[[2]]
  - best hyperparameter combo (lambda1, ratio, lambda2)
}
