% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_spec.R
\name{get_common_cols}
\alias{get_common_cols}
\title{Get Common Effect Columns}
\usage{
get_common_cols(spec)
}
\arguments{
\item{spec}{A matrix_spec object}
}
\value{
Integer vector c(start, end) or NULL
}
\description{
Get Common Effect Columns
}
