% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_spec.R
\name{get_common_tvp_cols}
\alias{get_common_tvp_cols}
\title{Get Common TVP Columns for a Period}
\usage{
get_common_tvp_cols(spec, period)
}
\arguments{
\item{spec}{A matrix_spec object}

\item{period}{Integer. Period index (1 to P)}
}
\value{
Integer vector c(start, end) or NULL
}
\description{
Get Common TVP Columns for a Period
}
