% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_n_periods}
\alias{get_n_periods}
\title{Get Number of Periods in TVP Model}
\usage{
get_n_periods(fit)
}
\arguments{
\item{fit}{A fitted multivar object}
}
\value{
Integer number of periods
}
\description{
Returns the number of time-varying periods in a TVP model.
Returns 1 for non-TVP models.
}
\examples{
\dontrun{
n_periods <- get_n_periods(fit)
}
}
