% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_spec.R
\name{get_subgrp_cols}
\alias{get_subgrp_cols}
\title{Get Subgroup Effect Columns}
\usage{
get_subgrp_cols(spec, subgroup)
}
\arguments{
\item{spec}{A matrix_spec object}

\item{subgroup}{Integer. Subgroup index (1 to S)}
}
\value{
Integer vector c(start, end) or NULL
}
\description{
Get Subgroup Effect Columns
}
