% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_spec.R
\name{get_subgrp_cols_for_subject}
\alias{get_subgrp_cols_for_subject}
\title{Get Subgroup Columns for a Subject}
\usage{
get_subgrp_cols_for_subject(spec, subject)
}
\arguments{
\item{spec}{A matrix_spec object}

\item{subject}{Integer. Subject index (1 to k)}
}
\value{
Integer vector c(start, end) or NULL
}
\description{
Get Subgroup Columns for a Subject
}
