% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_spec.R
\name{get_unique_cols}
\alias{get_unique_cols}
\title{Get Unique Effect Columns for a Subject}
\usage{
get_unique_cols(spec, subject)
}
\arguments{
\item{spec}{A matrix_spec object}

\item{subject}{Integer. Subject index (1 to k)}
}
\value{
Integer vector c(start, end) or NULL (for k=1 TVP)
}
\description{
Get Unique Effect Columns for a Subject
}
