% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_spec.R
\name{get_unique_tvp_cols}
\alias{get_unique_tvp_cols}
\title{Get Unique TVP Columns for a Subject and Period}
\usage{
get_unique_tvp_cols(spec, subject, period)
}
\arguments{
\item{spec}{A matrix_spec object}

\item{subject}{Integer. Subject index (1 to k)}

\item{period}{Integer. Period index (1 to P_i)}
}
\value{
Integer vector c(start, end) or NULL
}
\description{
Get Unique TVP Columns for a Subject and Period
}
