% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_na.R
\name{handle_na}
\alias{handle_na}
\title{Handle missing values
Adjust missing values in data frame (typically -99) to NA so they can be
properly handled in further operations. Note this operation inserts NA's into
the entire row in which there is a missing value.}
\usage{
handle_na(data, na_value, col)
}
\arguments{
\item{data}{data frame or data table with missing values}

\item{na_value}{The 'missing' value which will be translated to NA within
\code{data} (typically -99)}

\item{col}{Which column within \code{data} should be checked for missing values}
}
\description{
Handle missing values
Adjust missing values in data frame (typically -99) to NA so they can be
properly handled in further operations. Note this operation inserts NA's into
the entire row in which there is a missing value.
}
