% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.sample.profile.R
\name{k.sample.profile}
\alias{k.sample.profile}
\title{Performs profile anaylsis on multivsriate data with k different groups}
\usage{
k.sample.profile(Data, Test = "P", K = 1)
}
\arguments{
\item{Data}{Dataframe with first column as category and the rest as mesured varaibles}

\item{K}{Number of categories in data}
}
\value{
MANOVA summary for desired profile test
}
\description{
Performs profile anaylsis on multivsriate data with k different groups
}
\examples{
one.sample.hotelling(data, Test = "p", K = 3)

}
