% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maity_debiasing.R
\name{maity_debiasing}
\alias{maity_debiasing}
\title{Maity et al. (2022) MrLasso-style initial estimation}
\description{
Provides debiased, robustly-aggregated, analytically-thresholded initial
estimates for adaptive LASSO weights. Replaces the default median-based
decomposition when \code{weightest = "maity"} in \code{\link{constructModel}}.
}
\references{
Maity, S., Sun, Y., & Banerjee, M. (2022). Meta-analysis of heterogeneous
data: integrative sparse regression in high-dimensions. \emph{JMLR}, 23, 1-50.
}
