% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_initial_coefs.R
\name{make_unit_weights}
\alias{make_unit_weights}
\title{Create unit weight matrices for standard (non-adaptive) LASSO}
\usage{
make_unit_weights(d, k, subgroup, subgroup_membership, tvp, Ak, breaks)
}
\description{
Create unit weight matrices for standard (non-adaptive) LASSO
}
\keyword{internal}
