% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{monthlyClimatology}
\alias{monthlyClimatology}
\title{Calculate monthly climatology}
\usage{
monthlyClimatology(d, climatologyYears, var)
}
\arguments{
\item{d}{a data.frame containing at least a column named month and one other variable}

\item{climatologyYears}{a vector of length two indicating the range of years
to calculate the climatology}

\item{var}{name of variable of which anomaly should be calculated (present in `d``)}
}
\value{
the results of aggregate
}
\description{
This function calculates the monthly climatology
}
\seealso{
Other anomaly: 
\code{\link{annualAnomaly}()},
\code{\link{annualNormalizedAnomaly}()},
\code{\link{monthlyAnomaly}()},
\code{\link{monthlyNormalizedAnomaly}()},
\code{\link{monthlyStandardDeviation}()}
}
\author{
Chantelle Layton, Emily Chisholm
}
\concept{anomaly}
