% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{monthlyNormalizedAnomaly}
\alias{monthlyNormalizedAnomaly}
\title{Caclulate monthly normalized anomaly values}
\usage{
monthlyNormalizedAnomaly(d, climatologyYears, var)
}
\arguments{
\item{d}{a data.frame containing at least a column named year, month, and one other variable}

\item{climatologyYears}{a vector of length two indicating the range of years
to calculate the climatology.}

\item{var}{name of variable of which anomaly should be calculated (present in `d``)}
}
\value{
a data.frame containing columns \code{year}, \code{month}, and \code{normalizedAnomaly}
}
\description{
This function calculates the monthly normalized anomaly.
}
\details{
To calculated the normalized anomaly, two climatology values are required, those being the
monthly and standard deviation. The normalized anomaly values are calculated
by subtracting the monthly climatology value from the monthly value, and the dividing it
by the monthly climatology standard deviation.
}
\seealso{
Other anomaly: 
\code{\link{annualAnomaly}()},
\code{\link{annualNormalizedAnomaly}()},
\code{\link{monthlyAnomaly}()},
\code{\link{monthlyClimatology}()},
\code{\link{monthlyStandardDeviation}()}
}
\author{
Chantelle Layton, Emily Chisholm
}
\concept{anomaly}
