% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{multivar_fit_methods}
\alias{multivar_fit_methods}
\alias{print.multivar_fit}
\alias{summary.multivar_fit}
\alias{plot.multivar_fit}
\alias{coef.multivar_fit}
\title{S3 Methods for multivar_fit Objects}
\usage{
print.multivar_fit(x, ...)

summary.multivar_fit(object, ...)

\method{plot}{multivar_fit}(x, type = c("dynamics", "prevalence"), ...)

\method{coef}{multivar_fit}(
  object,
  type = c("total", "common", "unique"),
  subject = NULL,
  period = NULL,
  ...
)
}
\arguments{
\item{x}{A multivar_fit object (output from cv.multivar())}

\item{...}{Additional arguments (currently unused)}

\item{object}{A multivar_fit object (output from cv.multivar())}
}
\description{
Standard S3 methods for fitted multivar models
}
