% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar_sim.R
\name{multivar_sim}
\alias{multivar_sim}
\title{Simulate multivar data.}
\usage{
multivar_sim(k, d, n, prop_fill_com, prop_fill_ind, lb, ub, sigma)
}
\arguments{
\item{k}{Integer. The number of individuals (or datsets) to be generated.}

\item{d}{Integer. The number of variables per dataset. For now this will be constant across individuals.}

\item{n}{Integer. The time series length.}

\item{prop_fill_com}{Numeric. The proportion of nonzero paths in the common transition matrix.}

\item{prop_fill_ind}{Numeric. The proportion of nonzero unique (not in the common transition matrix or transition matrix of other individuals) paths in each individual transition matrix.}

\item{lb}{Numeric. The upper bound for individual elements of the transition matrices.}

\item{ub}{Numeric. The lower bound for individual elements of the transition matrices.}

\item{sigma}{Matrix. The (population) innovation covariance matrix.}
}
\description{
Simulate multivar data.
}
\examples{

k <- 3
d <- 5
n <- 100
prop_fill_com <- .2
prop_fill_ind <- .2
lb <- 0.1
ub <- 0.7
sigma <- diag(0.1,d)
data <- multivar_sim(k, d, n, prop_fill_com, prop_fill_ind, lb, ub,sigma)$data

}
\keyword{multivar}
\keyword{simulate}
\keyword{var}
