% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar_sim_breaks.R
\name{multivar_sim_breaks}
\alias{multivar_sim_breaks}
\title{Simulate multivar data with time-varying dynamics according to a latent growth curve.}
\usage{
multivar_sim_breaks(
  k,
  d,
  n,
  prop_fill_com,
  prop_fill_ind,
  prop_fill_com_growth,
  prop_fill_ind_growth,
  offset = 0,
  lb,
  ub,
  lb_int,
  ub_int,
  lb_slope,
  ub_slope,
  sigma,
  diag = FALSE,
  iters = 1,
  intercept = NULL
)
}
\arguments{
\item{k}{Integer. The number of individuals (or datasets) to be generated.}

\item{d}{Integer. The number of variables per dataset. For now this will be constant across individuals.}

\item{n}{Integer. The time series length.}

\item{prop_fill_com}{Numeric. The proportion of nonzero paths in the common transition matrix.}

\item{prop_fill_ind}{Numeric. The proportion of nonzero unique (not in the common transition matrix or transition matrix of other individuals) paths in each individual transition matrix.}

\item{prop_fill_com_growth}{Numeric. The proportion of nonzero time-varying paths in the common transition matrix.}

\item{prop_fill_ind_growth}{Numeric. The proportion of nonzero time-varying unique (not in the common transition matrix or transition matrix of other individuals) paths in each individual transition matrix.}

\item{offset}{Numeric. An offset dictating when dynamics should transition from time-invariant to time-varying.}

\item{lb}{Numeric. The upper bound for individual elements of the transition matrices.}

\item{ub}{Numeric. The lower bound for individual elements of the transition matrices.}

\item{sigma}{Matrix. The (population) innovation covariance matrix.}

\item{diag}{Logical. Default is FALSE. Should diagonal elements be filled first for common elements.}

\item{intercept}{List. Default is NULL. A list of length K containing numeric vectors of length d representing the intercept values. If NULL, intercepts are set to 0.}
}
\description{
Simulate multivar data with time-varying dynamics according to a latent growth curve.
}
\keyword{multivar}
\keyword{simulate}
\keyword{var}
