% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.sample.hotelling.R
\name{one.sample.hotelling}
\alias{one.sample.hotelling}
\title{Performs a one sample hotelling t^2-test for given hypothesis}
\usage{
one.sample.hotelling(Data, mu.0)
}
\arguments{
\item{mu.0}{Hypothesized values of means}

\item{data}{Matrix of vectors (variables) you want to test means of}
}
\value{
F-stat and p-value for hotelling t^2 test
}
\description{
Performs a one sample hotelling t^2-test for given hypothesis
}
\examples{
one.sample.hotelling(iris[,1:4], c(5,3,4,1))

}
