% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{PCA}
\alias{PCA}
\title{Principal component analysis for given set of variables
Performs a principal components analysis using the stats package, on all
variables provided in \code{data} with the exception of \code{year}. Converts to
data.table and removes duplicate or incomplete rows before processing.}
\usage{
PCA(data)
}
\arguments{
\item{data}{a datatable or dataframe with varibles to be analyzed for PCA}
}
\value{

}
\description{
Principal component analysis for given set of variables
Performs a principal components analysis using the stats package, on all
variables provided in \code{data} with the exception of \code{year}. Converts to
data.table and removes duplicate or incomplete rows before processing.
}
\author{
Daniel Reed, Emily Chisholm
}
