% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cv_lambda_grid.R
\name{plot_cv_lambda_grid}
\alias{plot_cv_lambda_grid}
\title{Plot CV error over (lambda1, lambda2)}
\usage{
plot_cv_lambda_grid(fit, object = NULL, log_scale = TRUE, show_best = TRUE)
}
\arguments{
\item{fit}{result from cv.multivar(...)}

\item{object}{OPTIONAL; only used if fit does NOT contain hyperparams or obj}

\item{log_scale}{logical; plot on log10 scale?}

\item{show_best}{logical; mark best pair?}
}
\description{
Plot CV error over (lambda1, lambda2)
}
