% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{plot_region}
\alias{plot_region}
\title{Plot a data region \link{UNDER CONSTRUCTION}}
\usage{
plot_region(name, longitudelim, latitudelim, plot = TRUE)
}
\arguments{
\item{name}{name of the region to visualize}

\item{longitudelim}{optional vector of two numbers describing longitude
limits of plotting window (passed to /code{/link{oce::mapPlot}}), if no
value is given uses a maritimes centred default}

\item{latitudelim}{optional vector of two numbers describing latitude limits
of plotting window (passed to /code{/link{oce::mapPlot}}), if no value is
given uses a maritimes centred default}

\item{plot}{Logical value, if FALSE, will output data frame of coordinates
which can be plotted by user}
}
\description{
Plot a data region \link{UNDER CONSTRUCTION}
}
