% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powergraph.R
\name{powergraph}
\alias{powergraph}
\title{Adds two to any number}
\usage{
powergraph(n, d, alpha, alt)
}
\arguments{
\item{x}{A number}
}
\value{
A number
}
\description{
Adds two to any number
}
\examples{
add_two(3)

}
