% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_spec.R
\name{print.matrix_spec}
\alias{print.matrix_spec}
\title{Print method for matrix_spec}
\usage{
print.matrix_spec(x, ...)
}
\arguments{
\item{x}{A matrix_spec object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for matrix_spec
}
