% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query azmpdata Data products}
\usage{
query(search_key, location = "title")
}
\arguments{
\item{search_key}{Keyword(s) to search for in azmpdata, if using multiple keywords, provide vector of character strings}

\item{location}{Where to search for keyword(s), options are 'title', or 'metadata'}
}
\description{
Search for matches to search term(s) within azmpdata package data objects
}
\note{
Leave search_key blank \code{''} to return all available datasets
}
\examples{

query(search_key = c('broadscale'), location = 'title')



}
