% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarObjectClass.R
\docType{methods}
\name{show.multivar}
\alias{show.multivar}
\alias{show,multivar-method}
\title{Default show method for an object of class multivar}
\usage{
\S4method{show}{multivar}(object)
}
\arguments{
\item{object}{\code{multivar} object created from \code{ConstructModel}}
}
\value{
Displays the following information about the multivar object:
\itemize{
\item{To do.}
}
}
\description{
Default show method for an object of class multivar
}
\seealso{
\code{\link{constructModel}}
}
