% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maity_debiasing.R
\name{soft_threshold_matrix}
\alias{soft_threshold_matrix}
\title{Element-wise soft thresholding of a matrix}
\usage{
soft_threshold_matrix(mat, threshold)
}
\arguments{
\item{mat}{Matrix to threshold}

\item{threshold}{Threshold value}
}
\value{
Thresholded matrix
}
\description{
Element-wise soft thresholding of a matrix
}
\keyword{internal}
