% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvp_sim.R
\name{tvp_sim}
\alias{tvp_sim}
\title{Simulate a time-varying VAR model.}
\usage{
tvp_sim(n, sigma, mat_total, mat_tvp, int = 0)
}
\arguments{
\item{n}{Integer. The time series length.}

\item{sigma}{Matrix. Innovation covariance matrix.}

\item{mat_total}{Matrix. Transition matrix.}

\item{mat_tvp}{A logical matrix indicating which parameters are time-varying.}
}
\description{
Simulate a time-varying VAR model.
}
\keyword{simulate}
\keyword{time-varying}
\keyword{var}
