% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_spec.R
\name{validate_matrix_spec}
\alias{validate_matrix_spec}
\title{Validate Matrix Specification}
\usage{
validate_matrix_spec(spec)
}
\arguments{
\item{spec}{A matrix_spec object}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
Checks internal consistency of a matrix_spec object.
}
