% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_sim.R
\name{var_sim}
\alias{var_sim}
\title{Simulate a stationary Vector Autoregressive (VAR) time series.}
\usage{
var_sim(n, phi, sigma, burn = 500)
}
\arguments{
\item{n}{An integer giving the number of timepoints.}

\item{phi}{A d x d transition matrix.}

\item{sigma}{A d x d innovation covariance matrix.}

\item{bubrn}{Burn-in period to be discarded. Default is 500.}
}
\description{
Simulate a stationary Vector Autoregressive (VAR) time series.
}
\examples{

theta    <- diag(c(.7,.8,.9,.6,.7,.9))
data     <- var_sim(100, theta, diag(.1,6))
datalag  <- embed(data, 2)
b        <- datalag[,1:6]
A        <- datalag[,7:12]
A_est    <- fista_sparse(A, b, 1, theta, niter = 10, backtrack = TRUE)$out.x
var_forecast(t(b), 2, A_est)

}
\keyword{simulate}
\keyword{var}
