% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_scenario_table.R
\name{build_scenario_table}
\alias{build_scenario_table}
\title{Reconstruct scenario-level multipliers used to build W}
\usage{
build_scenario_table(object)
}
\arguments{
\item{object}{multivar object built using \code{ConstructModel}.}
}
\description{
This matches the logic in est_base_weight_mat(): we expand
ratios_unique, and (optionally) ratios_subgroup and ratios_unique_tvp into a list
of S scenarios, where S = dim(W)[3].
}
