% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarObjectClass.R
\name{constructModel}
\alias{constructModel}
\title{Construct an object of class multivar}
\usage{
constructModel(data = NULL, lag = 1, struct = "Sparse",
  horizon = 1, t1 = NULL, t2 = NULL, lambdas = NULL, tol = 1e-04,
  window = 1, standardize = T, sep = NULL, header = NULL)
}
\arguments{
\item{data}{List. A list (length = k) of T by d multivariate time series}

\item{lag}{Numeric. The VAR order. Default is 1.}

\item{struct}{Character. The choice of penalty structure can be "Sparse","Lowrank",or "SparseLowrank".}

\item{horizon}{Numeric. Desired forecast horizon. Default is 1.}

\item{t1}{Numeric. Index of time series in which to start cross validation. If NULL, default is floor(nrow(n)/3) where nk is the time series length for individual k.}

\item{t2}{Numeric. Index of times series in which to end cross validation. If NULL, default is floor(2*nrow(n)/3) where nk is the time series length for individual k.}

\item{lambdas}{List. User-supplied penalty parameters.}

\item{tol}{Numeric. Optimization tolerance (default 1e-4)}

\item{window}{Numeric. Size of rolling window.  If set to 0 an expanding window will be used.}

\item{standardize}{Logical. Default is true. Whether to standardize the individual data.}
}
\description{
Construct an object of class multivar
}
\details{
The choices for "struct" are as follows
\itemize{
\item{  "Sparse" (Sparse)}
\item{  "Lowrank" (Low-Rank)} 
\item{  "SparseLowrank" (Sparse and Low-Rank)} 
}
}
\examples{
# multivar Example
}
