% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarObjectClass.R
\docType{methods}
\name{cv.multivar}
\alias{cv.multivar}
\alias{cv.multivar,multivar-method}
\title{Cross Validation for multivar}
\usage{
cv.multivar(object)
}
\arguments{
\item{object}{multivar object built using \code{ConstructModel}.}
}
\value{
An object of class \code{multivar.results}.
}
\description{
Cross Validation for multivar
}
\details{
The main function of the multivar package. Performs cross validation to select penalty parameters over a training sample and evaluates them over a test set. Creates an object of class \code{multivar.results}
}
\examples{
 # Add example
}
\seealso{
\code{\link{constructModel}}, \code{\link{multivar.results}},\code{\link{multivar.est}}
}
