% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\docType{package}
\name{multivar-package}
\alias{multivar}
\alias{multivar-package}
\title{Penalized Estimation and Forecasting of Multiple Subject VAR Models}
\description{
\pkg{multivar} is a package for estimating and forecasting penalized
Vector Autoregression (VAR) models from multiple subjects..
}
\author{
\strong{Maintainer}: Zachary Fisher \email{fish.zachary@gmail.com}

Authors:
\itemize{
  \item Younghoon Kim
  \item Vladas Pipiras
}

}
