% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar.R
\docType{package}
\name{multivar-package}
\alias{multivar}
\alias{multivar-package}
\title{Penalized Estimation of Multiple Subject Vector Autoregressive (multi-VAR) Models}
\description{
\pkg{multivar} is an R package for simulating, estimating and forecasting stationary Vector Autoregressive (VAR) models for multiple subject data using the penalized multi-VAR framework.
}
\author{
\strong{Maintainer}: Zachary Fisher \email{fish.zachary@gmail.com}

Other contributors:
\itemize{
  \item Younghoon Kim [contributor]
  \item Vladas Pipiras [contributor]
}

}
