% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivar-package.R, R/multivar.R
\docType{package}
\name{multivar-package}
\alias{multivar-package}
\title{multivar: Penalized Estimation of Multiple-Subject Vector Autoregressive (Multi-VAR) Models}
\description{
Functions for simulating, estimating and forecasting Vector Autoregressive (VAR) models for multiple-subject data using structured penalization.

\pkg{multivar} is an R package for simulating, estimating and forecasting stationary Vector Autoregressive (VAR) models for multiple subject data using the penalized multi-VAR framework.
}
\author{
\strong{Maintainer}: Zachary Fisher \email{fish.zachary@gmail.com}

Authors:
\itemize{
  \item Christopher Crawford
}

Other contributors:
\itemize{
  \item Younghoon Kim [contributor]
  \item Vladas Pipiras [contributor]
}

}
