% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{PCA}
\alias{PCA}
\title{Principal component analysis for given set of variables}
\usage{
PCA(data)
}
\arguments{
\item{data}{a datatable or dataframe with varibles to be analyzed for PCA}
}
\value{

}
\description{
Principal component analysis for given set of variables
}
