% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transition_mat.R
\name{plot_transition_mat}
\alias{plot_transition_mat}
\title{Plot arbitrary transition matrix.}
\usage{
plot_transition_mat(x, title = NULL, subtitle = NULL, ub = 1, lb = -1)
}
\arguments{
\item{x}{Matrix. An arbitrary transition matrix.}

\item{title}{Character. A title for the plot.}

\item{subtitle}{Character. A subtitle for the plot.}

\item{ub}{Numeric. Upper bound on coefficient values for heatmap index. Default is 1.}

\item{lb}{Numeric. Lower bound on coefficient values for heatmap index. Default is -1.}
}
\description{
Plot arbitrary transition matrix.
}
\examples{

plot_transition_mat(matrix(rnorm(25),5,5), title= "Example")

}
\keyword{lot}
\keyword{multivar}
\keyword{var}
